Switch from okhttp WS client to java-websocket-client.
This may not last, but there's some issues with the okhttp client (or the implementation if it), where it gets in a disconnected state and stays there.

This particular client has state methods (isOpen, isClosed, etc) - which makes it easier to maintain the connection in our client (IMO). And allows us to re-use the client (reconnect) instead of building a new one in some cases.